mardizzone/pos-649 (#121)
* dev: add: pos-649: add first version of integration test

* dev: chg: pos-649: chg path

* dev: chg: pos-649: chg path/2

* dev: chg: pos-649: chg path/3

* dev: chg: pos-649: chg path/4

* dev: chg: pos-649: chg path/5

* dev: chg: pos-649: fix ref path and some env vars

* dev: fix monitor service and rename duplicated loadConfig method

* dev: chg: use express-cli remotely into CI executions

* dev: chg: fix sed paths

* dev: chg: add more .env vars / install aws-cli instructions for linux

* dev: chg: update aws-cli in case already installed

* dev: fix: install node and dependencies

* dev: add: remove keypair after execution / remove logs

* dev: add: strict permissions on key pair

* dev: rm: delete matic-cli-config.yml previously used for ci

* dev: rm: add write permissions to keypair

* dev: add: use aws sso for terraform-aws operations

* dev: fix: typo

+++ src/express/commands/destroy.js
@@ -10,9 +10,10 @@ export async function terraformDestroy() {
             ...process.env,
         }
     });
-
-    let devnetId = getDevnetId()
-    console.log(`📍Deleting ./deployments/devnet-${devnetId} folder...`)
-    shell.exec(`cd ../../ && rm -rf ./deployments/devnet-${devnetId}`)
-    console.log(`📍devnet-${devnetId} successfully destroyed!`)
+    if (shell.error() == null) {
+        let devnetId = getDevnetId()
+        console.log(`📍Deleting ./deployments/devnet-${devnetId} folder...`)
+        shell.exec(`cd ../../ && rm -rf ./deployments/devnet-${devnetId}`)
+        console.log(`📍devnet-${devnetId} successfully destroyed!`)
+    }
 }
